Move ut_cs2cep.py to ellipse_params.py, and add matlab help as temp docstring.
